Generalise tween function
