+++ src/containers/Start.js
@@ -7,7 +7,7 @@ import { get } from 'lodash';
 import cx from 'classnames';
 import { compose, mapProps } from 'recompose';
 import { Button, Icon } from 'network-canvas-ui';
-import TweenFrom from '../behaviours/Tween/From';
+import Tweened from '../behaviours/Tweened';
 import { actionCreators as protocolsActions } from '../ducks/modules/protocols';
 import { ProtocolCard } from '../components/Start';
 import networkCanvasBrand from '../images/network-canvas-brand.svg';
@@ -23,7 +23,7 @@ const transitionStyles = (componentName) =>
     })
   );
 
-const TweenedProtocolCard = TweenFrom(ProtocolCard);
+const TweenedProtocolCard = Tweened(ProtocolCard);
 
 class Start extends PureComponent {
   static propTypes = {
